requirements.txt
asgiref==3.8.1
cassandra-driver==3.29.2
click==8.1.7
Django==5.1.1
django-cassandra-engine==1.9.0
geomet==0.2.1.post1
PyYAML==6.0.2
scylla-driver==3.26.9
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
